Summary: Simplified the parsing process by incorporating the parseArguments().
Changed vm to be a smart pointer. But due to these changes had to rewrite quite a few lines of code. Now the code works as intended.

Also tried incorporating the flag system into parser but since flowControl is dependent on parser had to abandon the idea. Perhaps I will then separate the flow control and parser.