Summary: Added URL argument
Added URL argument so that user can directly send a link to fetch. The program will handle it as if the user specified the data etc. The advantage is that the user can link any URL and the program will try and fetch it. But it may cause unintended behavior so the user should be warned.

Updated descriptions